requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
blinker==1.8.2
    # via flask
cffi==1.17.1
    # via gevent
click==8.1.7
    # via flask
colorama==0.4.6
    # via click
colorzero==2.0
    # via gpiozero
flask==3.0.3
    # via rotary-phone-audio-guestbook (pyproject.toml)
gevent==24.11.1
    # via rotary-phone-audio-guestbook (pyproject.toml)
gpiozero==2.0.1
    # via rotary-phone-audio-guestbook (pyproject.toml)
greenlet==3.1.1
    # via gevent
gunicorn==23.0.0
    # via rotary-phone-audio-guestbook (pyproject.toml)
importlib-metadata==4.13.0
    # via
    #   flask
    #   gpiozero
importlib-resources==5.13.0
    # via gpiozero
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
packaging==24.1
    # via gunicorn
psutil==7.0.0
    # via rotary-phone-audio-guestbook (pyproject.toml)
pycparser==2.22
    # via cffi
ruamel-yaml==0.18.6
    # via rotary-phone-audio-guestbook (pyproject.toml)
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
setuptools==74.1.2
    # via
    #   colorzero
    #   zope-event
    #   zope-interface
werkzeug==3.0.4
    # via flask
zipp==3.20.2
    # via
    #   importlib-metadata
    #   importlib-resources
zope-event==5.0
    # via gevent
zope-interface==7.2
    # via gevent
pydub
pyaudio
